# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Babylon Nexus API
  version: 1.0.0
  description: |
    ## Overview
    The Babylon Nexus API manages construction project workflows including authentication,
    project management, invoice lifecycle, and participant coordination.

    ## Invoice Lifecycle
    Invoices follow this sequential workflow:
    ```
    Pending → Approved → Paid → Received
                    ↘ Rejected (from Pending only)
    ```
    The approver is automatically resolved at submission time based on the submitter's role
    and which roles are present on the project (`approverRole` stored on the invoice).

    ## Authentication
    All protected routes require a Bearer JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens expire in **10 minutes**. Use `/auth/refresh` to obtain a new one.
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from `/auth/login` or `/auth/refresh`

  # ─────────────────────────────────────────────
  # Reusable Schemas
  # ─────────────────────────────────────────────
  schemas:
...
    Project:
      type: object
      description: Represents a construction project (maps to PROJECT table)
      properties:
        id:
          type: string
          example: "proj_123e4567-e89b-12d3-a456-426614174000"
        location:
          type: string
          example: "2-4 Mintaro Ave, Strathfield 2135 (Lot 1, DP: 954705)"
        council:
          type: string
          example: "Strathfield"
        ownerId:
          type: string
          description: USER.id of the project owner
          example: "user_owner123"
        builderId:
          type: string
          description: USER.id of the builder
          example: "user_builder123"
        pmId:
          type: string
          description: USER.id of the project manager
          example: "user_pm123"
        status:
          type: string
          example: "90% Complete"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Descriptive error message"
...
  responses:
    Unauthorized:
      description: Invalid or missing access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication Required"
    BadRequest:
      description: Invalid or missing fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /projects:
    get:
      tags:
        - Multiple Roles
      summary: Get all associated projects
      description: |
        Returns all projects associated with the authenticated user. Results are filtered by role:

        | Role | Projects returned |
        |------|------------------|
        | Project Manager | Projects where user is the PM (`PROJECT.pmId`) |
        | Builder | Projects where user is the builder (`PROJECT.builderId`) |
        | Owner | Projects where user is the owner (`PROJECT.ownerId`) |
        | Sub Contractor / Consultant | Projects in `PROJECT_PARTICIPANTS` |
      responses:
        '200':
          description: List of associated projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer
                    description: Total number of projects
                    example: 15
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /project:
    post:
      tags:
        - Multiple Roles
      summary: Create a new project
      description: |
        Creates a new project. **Any authenticated user can create a project** — there is no role restriction.
        The creator is automatically added as a participant with the specified `creatorRole`.
        The project starts with `status: "Pending"` and must be approved by Admin before it becomes `Active`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, council, creatorRole]
              properties:
                name:
                  type: string
                  description: Optional project name
                  example: "Strathfield Residential Build"
                location:
                  type: string
                  example: "2-4 Mintaro Ave, Strathfield 2135"
                council:
                  type: string
                  example: "Strathfield"
                creatorRole:
                  type: string
                  description: The role the creator will hold on this project
                  enum: [PM, Subbie, Owner, Builder, Consultant, Financier, VIP, Observer]
                  example: "PM"
      responses:
        '200':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  projectId:
                    type: string
                    example: "proj_123e4567-e89b-12d3-a456-426614174000"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...